+++ src/ui/marker.js
@@ -1,17 +1,17 @@
 // @flow
 
-import DOM from '../util/dom';
-import window from '../util/window';
-import LngLat from '../geo/lng_lat';
+import DOM from '../util/dom.js';
+import window from '../util/window.js';
+import LngLat from '../geo/lng_lat.js';
 import Point from '@mapbox/point-geometry';
-import smartWrap from '../util/smart_wrap';
-import {bindAll, extend} from '../util/util';
-import {type Anchor, anchorTranslate, applyAnchorClass} from './anchor';
-import {Event, Evented} from '../util/evented';
-import type Map from './map';
-import type Popup from './popup';
-import type {LngLatLike} from "../geo/lng_lat";
-import type {MapMouseEvent, MapTouchEvent} from './events';
+import smartWrap from '../util/smart_wrap.js';
+import {bindAll, extend} from '../util/util.js';
+import {type Anchor, anchorTranslate, applyAnchorClass} from './anchor.js';
+import {Event, Evented} from '../util/evented.js';
+import type Map from './map.js';
+import type Popup from './popup.js';
+import type {LngLatLike} from "../geo/lng_lat.js";
+import type {MapMouseEvent, MapTouchEvent} from './events.js';
 import type {PointLike} from '@mapbox/point-geometry';
 
 type Options = {
@@ -21,10 +21,13 @@ type Options = {
     color?: string,
     scale?: number,
     draggable?: boolean,
+    clickTolerance?: number,
     rotation?: number,
     rotationAlignment?: string,
     pitchAlignment?: string
 };
+
+export const TERRAIN_OCCLUDED_OPACITY = 0.2;
 
 /**
  * Creates a marker component
@@ -36,12 +39,20 @@ type Options = {
  * @param {string} [options.color='#3FB1CE'] The color to use for the default marker if options.element is not provided. The default is light blue.
  * @param {number} [options.scale=1] The scale to use for the default marker if options.element is not provided. The default scale corresponds to a height of `41px` and a width of `27px`.
  * @param {boolean} [options.draggable=false] A boolean indicating whether or not a marker is able to be dragged to a new position on the map.
+ * @param {number} [options.clickTolerance=0] The max number of pixels a user can shift the mouse pointer during a click on the marker for it to be considered a valid click (as opposed to a marker drag). The default is to inherit map's clickTolerance.
  * @param {number} [options.rotation=0] The rotation angle of the marker in degrees, relative to its respective `rotationAlignment` setting. A positive value will rotate the marker clockwise.
  * @param {string} [options.pitchAlignment='auto'] `map` aligns the `Marker` to the plane of the map. `viewport` aligns the `Marker` to the plane of the viewport. `auto` automatically matches the value of `rotationAlignment`.
  * @param {string} [options.rotationAlignment='auto'] `map` aligns the `Marker`'s rotation relative to the map, maintaining a bearing as the map rotates. `viewport` aligns the `Marker`'s rotation relative to the viewport, agnostic to map rotations. `auto` is equivalent to `viewport`.
  * @example
  * var marker = new mapboxgl.Marker()
  *   .setLngLat([30.5, 50.5])
+ *   .addTo(map);
+ * @example
+ * // Set options
+ * var marker = new mapboxgl.Marker({
+ *     color: "#FFFFFF",
+ *     draggable: true
+ *   }).setLngLat([30.5, 50.5])
  *   .addTo(map);
  * @see [Add custom icons with Markers](https://www.mapbox.com/mapbox-gl-js/example/custom-marker-icons/)
  * @see [Create a draggable Marker](https://www.mapbox.com/mapbox-gl-js/example/drag-a-marker/)
@@ -58,12 +69,16 @@ export default class Marker extends Evented {
     _scale: number;
     _defaultMarker: boolean;
     _draggable: boolean;
+    _clickTolerance: number;
+    _isDragging: boolean;
     _state: 'inactive' | 'pending' | 'active'; // used for handling drag events
-    _positionDelta: ?number;
+    _positionDelta: ?Point;
+    _pointerdownPos: ?Point;
     _rotation: number;
     _pitchAlignment: string;
     _rotationAlignment: string;
     _originalTabIndex: ?string; // original tabindex of _element
+    _fadeTimer: ?TimeoutID;
 
     constructor(options?: Options, legacyOptions?: Options) {
         super();
@@ -79,13 +94,16 @@ export default class Marker extends Evented {
             '_onUp',
             '_addDragHandler',
             '_onMapClick',
-            '_onKeyPress'
+            '_onKeyPress',
+            '_clearFadeTimer'
         ], this);
 
         this._anchor = options && options.anchor || 'center';
         this._color = options && options.color || '#3FB1CE';
         this._scale = options && options.scale || 1;
         this._draggable = options && options.draggable || false;
+        this._clickTolerance = options && options.clickTolerance || 0;
+        this._isDragging = false;
         this._state = 'inactive';
         this._rotation = options && options.rotation || 0;
         this._rotationAlignment = options && options.rotationAlignment || 'auto';
@@ -213,12 +231,6 @@ export default class Marker extends Evented {
             // prevent focusing on click
             e.preventDefault();
         });
-        this._element.addEventListener('focus', () => {
-            // revert the default scrolling action of the container
-            const el = this._map.getContainer();
-            el.scrollTop = 0;
-            el.scrollLeft = 0;
-        });
         applyAnchorClass(this._element, this._anchor, 'marker');
 
         this._popup = null;
@@ -239,6 +251,8 @@ export default class Marker extends Evented {
         map.getCanvasContainer().appendChild(this._element);
         map.on('move', this._update);
         map.on('moveend', this._update);
+        map.on('remove', this._clearFadeTimer);
+        map._addMarker(this);
         this.setDraggable(this._draggable);
         this._update();
 
@@ -268,8 +282,11 @@ export default class Marker extends Evented {
             this._map.off('touchend', this._onUp);
             this._map.off('mousemove', this._onMove);
             this._map.off('touchmove', this._onMove);
+            this._map.off('remove', this._clearFadeTimer);
+            this._map._removeMarker(this);
             delete this._map;
         }
+        this._clearFadeTimer();
         DOM.remove(this._element);
         if (this._popup) this._popup.remove();
         return this;
@@ -431,6 +448,50 @@ export default class Marker extends Evented {
         return this;
     }
 
+    _evaluateOpacity() {
+        const position = this._pos ? this._pos.sub(this._transformedOffset()) : null;
+
+        if (!this._withinScreenBounds(position)) {
+            this._clearFadeTimer();
+            return;
+        }
+
+        const mapLocation = this._map.unproject(position);
+
+        let terrainOccluded = false;
+        if (this._map.getTerrain()) {
+            const camera = this._map.getFreeCameraOptions();
+            if (camera.position) {
+                const cameraPos = camera.position.toLngLat();
+                // the distance to the marker lat/lng + marker offset location
+                const offsetDistance = cameraPos.distanceTo(mapLocation);
+                const distance = cameraPos.distanceTo(this._lngLat);
+                terrainOccluded = offsetDistance < distance * 0.9;
+            }
+        }
+
+        const fogOpacity = this._map._queryFogOpacity(mapLocation);
+        const opacity = (1.0 - fogOpacity) * (terrainOccluded ? TERRAIN_OCCLUDED_OPACITY : 1.0);
+        this._element.style.opacity = `${opacity}`;
+        if (this._popup) this._popup._setOpacity(`${opacity}`);
+
+        this._fadeTimer = null;
+    }
+
+    _clearFadeTimer() {
+        if (this._fadeTimer) {
+            clearTimeout(this._fadeTimer);
+            this._fadeTimer = null;
+        }
+    }
+
+    _withinScreenBounds(position: ?Point): boolean {
+        const tr = this._map.transform;
+        return !!position &&
+            position.x >= 0 && position.x < tr.width &&
+            position.y >= 0 && position.y < tr.height;
+    }
+
     _update(e?: {type: 'move' | 'moveend'}) {
         if (!this._map) return;
 
@@ -438,7 +499,7 @@ export default class Marker extends Evented {
             this._lngLat = smartWrap(this._lngLat, this._pos, this._map.transform);
         }
 
-        this._pos = this._map.project(this._lngLat)._add(this._offset);
+        this._pos = this._map.project(this._lngLat)._add(this._transformedOffset());
 
         let rotation = "";
         if (this._rotationAlignment === "viewport" || this._rotationAlignment === "auto") {
@@ -461,7 +522,31 @@ export default class Marker extends Evented {
             this._pos = this._pos.round();
         }
 
-        DOM.setTransform(this._element, `${anchorTranslate[this._anchor]} translate(${this._pos.x}px, ${this._pos.y}px) ${pitch} ${rotation}`);
+        this._map._requestDomTask(() => {
+            if (!this._map) return;
+
+            if (this._element && this._pos && this._anchor) {
+                DOM.setTransform(this._element, `${anchorTranslate[this._anchor]} translate(${this._pos.x}px, ${this._pos.y}px) ${pitch} ${rotation}`);
+            }
+
+            if ((this._map.getTerrain() || this._map.getFog()) && !this._fadeTimer) {
+                this._fadeTimer = setTimeout(this._evaluateOpacity.bind(this), 60);
+            }
+        });
+    }
+
+    /**
+     * This is initially added to fix the behavior of default symbols only, in order
+     * to prevent any regression for custom symbols in client code.
+     * @private
+     */
+    _transformedOffset() {
+        if (!this._defaultMarker) return this._offset;
+        const tr = this._map.transform;
+        const offset = this._offset.mult(this._scale);
+        if (this._rotationAlignment === "map") offset._rotate(tr.angle);
+        if (this._pitchAlignment === "map") offset.y *= Math.cos(tr._pitch);
+        return offset;
     }
 
     /**
@@ -484,6 +569,12 @@ export default class Marker extends Evented {
     }
 
     _onMove(e: MapMouseEvent | MapTouchEvent) {
+        if (!this._isDragging) {
+            const clickTolerance = this._clickTolerance || this._map._clickTolerance;
+            this._isDragging = e.point.dist(this._pointerdownPos) >= clickTolerance;
+        }
+        if (!this._isDragging) return;
+
         this._pos = e.point.sub(this._positionDelta);
         this._lngLat = this._map.unproject(this._pos);
         this.setLngLat(this._lngLat);
@@ -524,6 +615,8 @@ export default class Marker extends Evented {
         // revert to normal pointer event handling
         this._element.style.pointerEvents = 'auto';
         this._positionDelta = null;
+        this._pointerdownPos = null;
+        this._isDragging = false;
         this._map.off('mousemove', this._onMove);
         this._map.off('touchmove', this._onMove);
 
@@ -554,7 +647,9 @@ export default class Marker extends Evented {
             // to calculate the new marker position.
             // If we don't do this, the marker 'jumps' to the click position
             // creating a jarring UX effect.
-            this._positionDelta = e.point.sub(this._pos).add(this._offset);
+            this._positionDelta = e.point.sub(this._pos).add(this._transformedOffset());
+
+            this._pointerdownPos = e.point;
 
             this._state = 'pending';
             this._map.on('mousemove', this._onMove);
@@ -597,7 +692,7 @@ export default class Marker extends Evented {
 
     /**
      * Sets the `rotation` property of the marker.
-     * @param {number} [rotation=0] The rotation angle of the marker (clockwise, in degrees), relative to its respective {@link Marker#rotationAlignment} setting.
+     * @param {number} [rotation=0] The rotation angle of the marker (clockwise, in degrees), relative to its respective {@link Marker#setRotationAlignment} setting.
      * @returns {Marker} `this`
      */
     setRotation(rotation: number) {